+++ packages/gatsby/src/redux/run-sift.js
@@ -21,11 +21,18 @@ const {
 
 const FAST_OPS = [
   `$eq`,
-  // "$lt",
+  `$ne`,
+  `$lt`,
   `$lte`,
-  // "$gt",
+  `$gt`,
   `$gte`,
+  `$in`,
+  `$nin`,
+  `$regex`, // Note: this includes $glob
 ]
+
+// More of a testing mechanic, to verify whether last runSift call used Sift
+let lastFilterUsedSift = false
 
 /**
  * Creates a key for one filterCache inside FiltersCache
@@ -390,6 +397,10 @@ const runFilterAndSort = (args: Object) => {
 
 exports.runSift = runFilterAndSort
 
+exports.didLastFilterUseSift = function _didLastFilterUseSift() {
+  return lastFilterUsedSift
+}
+
 /**
  * Applies filter. First through a simple approach, which is much faster than
  * running sift, but not as versatile and correct. If no nodes were found then
@@ -434,6 +445,8 @@ const applyFilters = (
   }
 
   const result = filterWithoutSift(filters, nodeTypeNames, filtersCache)
+
+  lastFilterUsedSift = false
   if (result) {
     if (stats) {
       stats.totalIndexHits++
@@ -443,6 +456,7 @@ const applyFilters = (
     }
     return result
   }
+  lastFilterUsedSift = true
 
   const siftResult = filterWithSift(
     filters,